+++ src/akkudoktoreos/prediction/elecpricetibber.py
@@ -299,10 +299,12 @@ class ElecPriceTibber(ElecPriceProvider):
         if not points:
             raise ValueError("Tibber response contains no price points")
 
+        self.log_charges_plan()
+
         series_data = pd.Series(dtype=float)
         for point in points:
             orig_datetime = to_datetime(point.startsAt, in_timezone=self.config.general.timezone)
-            series_data.at[orig_datetime] = point.total / 1000.0
+            series_data.at[orig_datetime] = self.apply_charges(point.total / 1000.0)
 
         return series_data.sort_index()